+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [PEP 440 about versioning](https://www.python.org/dev/pe
 - Add a `run_tests.py` file.
 - Allow using spaces around colon when naming a resource in a query.
 - Better exceptions
+- Manage slicing/list indexing in filters
 
 ## [0.0.1a1] - 2015-06-15
 - First working version